core_number was inspired by NetworkX but there are some efficiencies that we can apply. Here are the results of the new code (`core_number2` below):
```
julia> g = DiGraph(200_000, 5_000_000)                                                                                                       
{200000, 5000000} directed simple Int64 graph                                                                                                
                                                                                                                                             
julia> c = core_number(g);                                                                                                                   
                                                                                                                                             
julia> include("core_number.jl")                                                                                                             
core_number2 (generic function with 1 method)                                                                                                
                                                                                                                                             
julia> c2 = core_number2(g);                                                                                                                 
                                                                                                                                             
julia> c == c2                                                                                                                               
true                                              

julia> @benchmark core_number(g)                                                                                                             
BenchmarkTools.Trial:                                                                                                                        
  memory estimate:  1.53 GiB                                                                                                                 
  allocs estimate:  5625718                                                                                                                  
  --------------                                                                                                                             
  minimum time:     3.842 s (7.01% GC)
  median time:      3.851 s (9.58% GC)
  mean time:        3.851 s (9.58% GC)
  maximum time:     3.859 s (12.15% GC)
  --------------
  samples:          2
  evals/sample:     1


julia> @benchmark core_number2(g)
BenchmarkTools.Trial: 
  memory estimate:  925.09 MiB
  allocs estimate:  3520476
  --------------
  minimum time:     1.245 s (5.69% GC)
  median time:      1.255 s (5.61% GC)
  mean time:        1.258 s (5.64% GC)
  maximum time:     1.278 s (5.26% GC)
  --------------
  samples:          4
  evals/sample:     1
```