@sbromberger 2020-01-25
On second glance, I'm not sure this is algorithmically correct. Closing out for now.